chore: split off some general functions to separate file
